backend/requirements.txt
# Azure ML and OpenAI
openai>=1.12.0
azure-identity>=1.15.0

# HTTP and API
requests>=2.31.0

# Data validation
pydantic>=2.5.0

# Configuration
pyyaml>=6.0.1

# Utilities
python-dateutil>=2.8.2